==> Lessons/05_Part1_DataVisualization.Rmd <==
---
title: "5: Part 1 - Data Visualization Basics"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2023"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---
\listoffigures

## Objectives
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## Opening discussion
Effective data visualization depends on purposeful choices about graph types. The ideal graph type depends on the type of data and the message the visualizer desires to communicate. The best visualizations are clear and simple. A good resource for data visualization is [Data to Viz](https://www.data-to-viz.com/), which includes both a decision tree for visualization types and explanation pages for each type of data, including links to R resources to create them. Take a few minutes to explore this website.

## Set Up

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse);library(lubridate);library(here)
library(ggridges)
#install.packages("ggridges") 
#ridge plots help look at distribution of variable over time
here()

PeterPaul.chem.nutrients <- 
  read.csv(here("Data/Processed_KEY/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv"), stringsAsFactors = TRUE)
PeterPaul.chem.nutrients.gathered <-
  read.csv(here("Data/Processed_KEY/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv"), stringsAsFactors = TRUE)
EPAair <- read.csv(here("Data/Processed_KEY/EPAair_O3_PM25_NC1819_Processed.csv"), stringsAsFactors = TRUE)

#change date into a date object
EPAair$Date <- ymd(EPAair$Date)
PeterPaul.chem.nutrients$sampledate <- ymd(PeterPaul.chem.nutrients$sampledate)
PeterPaul.chem.nutrients.gathered$sampledate <- ymd(PeterPaul.chem.nutrients.gathered$sampledate)

```

## ggplot

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_density_ridges
* geom_point
* geom_errorbar
* geom_smooth - dataset you want to add a trendline to
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text - add text to your plot or change default of labels

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape - shape of point that you are using
* size - size of point you are using
* transparency - to points or bars 

### Plotting continuous variables over time: Scatterplot and Line Plot

Playing around with the r code chunk with fig.cap and changing the height

```{r, fig.height = 3, fig.width = 4, warning = FALSE, fig.cap="Scatterplot"}
# Scatterplot
#adding the fig.height = 3, fig.width = 4, warning = FALSE at the beginning of r code chunk means it is applied to every plot in the chunk

#x-axis = date and y-axis = ozone
#plot ozone as a point -->  geom_point()
ggplot(EPAair, aes(x = Date, y = Ozone)) + 
  geom_point()

#same plot as above, but create an object to assign ggplot to and then print the plot (helpful when you want to add multiple plots on same page or same window)
#store all of the points in an object you want
#note that the aes are outside of ggplot (compared to above)
O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot)

# Fix this code
#when you specify colors inside of aes you are adding another layer to plot, not changing color of points 
O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone, color = "blue"))
print(O3plot2)

#change color - must be OUTSIDE aes command
O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone), color = "blue")
print(O3plot2)

# Add additional variables
# How could you automatically assign a marker color to a variable? 

#asking r to automatically assign a shape that's not the default to year (different shapes depending on which year the data were recorded)
#similarly, color =  Site.Name means that it will automatically assign a color to a site name and give different colors for different site names
PMplot <- 
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year), color = Site.Name)) +
  geom_point()
print(PMplot)

# Separate plot with facets - going to use facet_wrap, removed color = and added facet_wrap
#creating a 4 row, 3 column of three different plots to show site names 
PMplot.faceted <-
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3)
print(PMplot.faceted)

# Filter dataset within plot building and facet by multiple variables
#you only want to plot a certain number or separate certain plots without removing them from dataset 
#use subset() - dataframe, which sites you want 
PMplot.faceted2 <- 
  ggplot(subset(EPAair, Site.Name == "Clemmons Middle" | Site.Name == "Leggett" |
                  Site.Name == "Bryson City"), 
         aes(x = Month, y = PM2.5)) + 
  geom_point() +
  facet_grid(Site.Name ~ Year) #Site.Name ~ Year --> separating the sites by years
print(PMplot.faceted2)

# Plot true time series with geom_line - want to connect points with a line (x-axis needs to change)
PMplot.line <- 
  ggplot(subset(EPAair, Site.Name == "Leggett"), #just looking at one site location
         aes(x = Date, y = PM2.5)) + #changing x-axis to date instead of month
  geom_line()
print(PMplot.line)

```


### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
#looking at relationship between irradiance Water and dissolved Oxygen
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO) #not a nice visualization of dataset - two outliers that stretch data

# Adjust axes
#add xlim and ylim to adjust axis (to exclude outliers)
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
  xlim(0, 250) + #max value you want is 250 
  ylim(0, 20) #want to go from 0-20
print(lightvsDOfixed)

# Depth in the fields of limnology and oceanography is on a reverse scale
#scale_y_reverse() - if you don't specify your range it will go from 0 onwards, this flips axis so max at bottom and 0 on top (important in limnology)
tempvsdepth <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse()
print(tempvsdepth)

#want to see relationship between tp_ug and tn_ug and check is relationship changes depending on depth so we're going to categorize the obs. by depth (surface = 1 color, etc.)
NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() +
  geom_smooth(method = lm) + #adds trendline method is the line type and it'll be lm (linear model)
  geom_abline(aes(slope = 16, intercept = 0)) #maybe you already know intercept and slope - geom_abline - specify slope and intercept 
print(NvsP)

```

### Plotting continuous vs. categorical variables
A traditional way to display summary statistics of continuous variables is a bar plot with error bars. Let's explore why this might not be the most effective way to display this type of data. Navigate to the Caveats page on Data to Viz (https://www.data-to-viz.com/caveats.html) and find the page that explores barplots and error bars. 

What might be more effective ways to display the information? Navigate to the boxplots page in the Caveats section to explore further.

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"? - categorize nutrients within boxplot using different colors / colors regards contour of shape, fill changes the fill of the boxplot (doesn't look nice) - color is always the way to go with boxplots
print(Nutrientplot3)

# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) + #categorizing the dotplot with different arguments for x and y-axes, binwidth
  geom_dotplot(aes(color = nutrient, fill = nutrient), binaxis = "y", #stacking observations over y, binwidth = 1 
               stackdir = "center", position = "dodge", dotsize = 2) #stacking direction in the center and position = "dodge" is bascially not overlaying of dots, and dotsize is size of dots printed
print(Nutrientplot4)

# Violin plot -- also shows you a rotated density plot 
Nutrientplot5 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #
print(Nutrientplot5)

# Frequency polygons
# Using a tidy dataset
# specify what you want on each axis and the color you want for the polygon
# not a good way to generate freq plot (long code and no legend)
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "darkred") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "blue") +
  geom_freqpoly(aes(x = no23), color = "royalblue") +
  geom_freqpoly(aes(x = po4), color = "red") 
print(Nutrientplot6)

# Using a gathered dataset
# much better way to create a freq plot
#color inside aesthetics gives you legend
Nutrientplot7 <-   
  ggplot(PeterPaul.chem.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)

# Frequency polygons have the risk of becoming spaghetti plots. 
# See <https://www.data-to-viz.com/caveat/spaghetti.html> for more info.

# Ridgeline plot
#use fill to fill below the curve of the frequency plot (using color would color the line not under the curve)
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(y = nutrient, x = concentration)) +
  geom_density_ridges(aes(fill = lakename), alpha = 0.5) #
print(Nutrientplot6)


```
